feat(go): move usage model and logic out of ee;
